requirements.txt
flask
tensorflow
opencv-python
Pillow
numpy